+++ MROZA-Show/MROZA-Show-web/src/test/java/com/mroza/utils/DatabaseUtils.java
@@ -87,7 +87,7 @@ public class DatabaseUtils {
         List<Program> programList = programsDao.selectAllPrograms();
         programsDao.deletePrograms(programList);
         List<KidTable> kidTableList = kidTablesDao.selectAllKidTables();
-        kidTableList.forEach((kidTable) -> kidTablesDao.deleteKidTableById(kidTable.getId()));
+        kidTableList.forEach((kidTable) -> kidTablesDao.deleteKidTableById(kidTable.getId().intValue()));
         List<Table> tableList = tablesDao.selectAllTables();
         tablesDao.deleteTables(tableList);
         utilsSqlSession.commit();
@@ -95,7 +95,7 @@ public class DatabaseUtils {
 
     public Kid setUpKid(String symbol)
     {
-        Kid kid = new Kid(-1, symbol, false, null, null);
+        Kid kid = new Kid(-1L, symbol, false, null, null);
         kidsDao.insertKid(kid);
         utilsSqlSession.commit();
         return kid;
@@ -115,14 +115,14 @@ public class DatabaseUtils {
 
     public Program setUpProgram(String symbol, String name, String description, Kid kid) {
         Program program = setUpProgram(symbol, name, description);
-        program.setKidId(kid.getId());
+        program.setKidId(kid.getId().intValue());
         programsDao.insertProgram(program);
         utilsSqlSession.commit();
         return program;
     }
 
     public Program setUpAssignedProgram(Kid kid, Program program) {
-        program.setKidId(kid.getId());
+        program.setKidId(kid.getId().intValue());
         programsDao.insertProgram(program);
         utilsSqlSession.commit();
         return program;
@@ -154,12 +154,12 @@ public class DatabaseUtils {
 
     public TableRow setUpRowWithFields(String rowName,int orderNumber, int generalizationNumber, int learningNumber, Table table){
         TableRow tableRow = new TableRow(rowName,orderNumber,learningNumber,generalizationNumber);
-        tableRow.setTableId(table.getId());
+        tableRow.setTableId(table.getId().intValue());
         tableRowsDao.insertTableRow(tableRow);
         utilsSqlSession.commit();
         List<TableField> tableFields = tableRow.getRowFields();
         for(TableField field : tableFields) {
-            field.setRowId(tableRow.getId());
+            field.setRowId(tableRow.getId().intValue());
             tableFieldsDao.insertTableField(field);
         }
         utilsSqlSession.commit();
@@ -167,7 +167,7 @@ public class DatabaseUtils {
     }
 
     public Period setUpPeriod(Date beginDate, Date endDate, Kid kid) {
-        Period period = new Period(beginDate, endDate, kid.getId());
+        Period period = new Period(beginDate, endDate, kid.getId().intValue());
         periodsDao.insertPeriod(period);
         utilsSqlSession.commit();
         return period;
@@ -190,8 +190,8 @@ public class DatabaseUtils {
         for(ResolvedField resolvedField : resolvedFields)
         {
             resolvedField.setValue("OK");
-            resolvedField.setKidTableId(resolvedField.getKidTable().getId());
-            resolvedField.setTableFieldId(resolvedField.getTableField().getId());
+            resolvedField.setKidTableId(resolvedField.getKidTable().getId().intValue());
+            resolvedField.setTableFieldId(resolvedField.getTableField().getId().intValue());
             resolvedFieldsDao.updateResolvedField(resolvedField);
             utilsSqlSession.commit();
         }
@@ -200,8 +200,8 @@ public class DatabaseUtils {
         kidTable.setCollectingGeneralization(true);
         kidTable.setCollectingGeneralization(true);
         kidTable.setLastModDate(new Date());
-        kidTable.setPeriodId(kidTable.getPeriod().getId());
-        kidTable.setTableId(kidTable.getTable().getId());
+        kidTable.setPeriodId(kidTable.getPeriod().getId().intValue());
+        kidTable.setTableId(kidTable.getTable().getId().intValue());
         kidTablesDao.updateKidTable(kidTable);
         utilsSqlSession.commit();
         return kidTable;
@@ -223,6 +223,6 @@ public class DatabaseUtils {
     }
 
     public List<Table> getTablesForProgram(Program program) {
-       return tablesDao.selectTablesByProgramIdWithEdgesRowsFields(program.getId());
+       return tablesDao.selectTablesByProgramIdWithEdgesRowsFields(program.getId().intValue());
     }
 }
